+++ packages/a11y/src/index.ts
@@ -1,1 +1,7 @@
+export * from './types'
 export * from './FocusScope'
+export * from './useFocus'
+export * from './useFocusAuto'
+export * from './useFocusContain'
+export * from './useFocusRestore'
+export * from './useFocusWithin'
