+++ packages/hooks/src/index.ts
@@ -17,7 +17,5 @@ export { useControlled } from './useControlled'
 export { useMediaQuery } from './useMediaQuery'
 export { useDebounce } from './useDebounce'
 export { useSearch } from './useSearch'
-export { useFocus } from './useFocus'
-export { useFocusWithin } from './useFocusWithin'
 export { useKeyboard } from './useKeyboard'
 export { useSSR } from './useSSR'
